fix issues

+++ binClassInitialSol.py
@@ -115,8 +115,8 @@ class BuildingPlate:
         y_start = y - shapes[0] + 1
         y_end = y + 1
         
-        # Use slicing to insert the binary part matrix
-        self.grid[y_start:y_end, x:x + shapes[1]] += partMatrix
+        # Use slicing to insert the binary part matrix (cast to uint8 to match grid dtype)
+        self.grid[y_start:y_end, x:x + shapes[1]] += partMatrix.astype(np.uint8)
         self.collision_backend.update_grid_region(self.grid_state, x, y, partMatrix, shapes)
         
         # Update enclosure box bounds incrementally - O(1)
